refact(cluster/manger): split into small files
