+++ pkg/cluster/manager/basic.go
@@ -1,0 +1,179 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package manager
+
+import (
+	"errors"
+
+	"github.com/joomcode/errorx"
+	perrs "github.com/pingcap/errors"
+	operator "github.com/pingcap/tiup/pkg/cluster/operation"
+	"github.com/pingcap/tiup/pkg/cluster/spec"
+	"github.com/pingcap/tiup/pkg/cluster/task"
+	"github.com/pingcap/tiup/pkg/logger/log"
+	"github.com/pingcap/tiup/pkg/meta"
+)
+
+// EnableCluster enable/disable the service in a cluster
+func (m *Manager) EnableCluster(name string, options operator.Options, isEnable bool) error {
+	if isEnable {
+		log.Infof("Enabling cluster %s...", name)
+	} else {
+		log.Infof("Disabling cluster %s...", name)
+	}
+
+	metadata, err := m.meta(name)
+	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	b := m.sshTaskBuilder(name, topo, base.User, options)
+
+	if isEnable {
+		b = b.Func("EnableCluster", func(ctx *task.Context) error {
+			return operator.Enable(ctx, topo, options, isEnable)
+		})
+	} else {
+		b = b.Func("DisableCluster", func(ctx *task.Context) error {
+			return operator.Enable(ctx, topo, options, isEnable)
+		})
+	}
+
+	t := b.Build()
+
+	if err := t.Execute(task.NewContext()); err != nil {
+		if errorx.Cast(err) != nil {
+			// FIXME: Map possible task errors and give suggestions.
+			return err
+		}
+		return perrs.Trace(err)
+	}
+
+	if isEnable {
+		log.Infof("Enabled cluster `%s` successfully", name)
+	} else {
+		log.Infof("Disabled cluster `%s` successfully", name)
+	}
+
+	return nil
+}
+
+// StartCluster start the cluster with specified name.
+func (m *Manager) StartCluster(name string, options operator.Options, fn ...func(b *task.Builder, metadata spec.Metadata)) error {
+	log.Infof("Starting cluster %s...", name)
+
+	metadata, err := m.meta(name)
+	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	tlsCfg, err := topo.TLSConfig(m.specManager.Path(name, spec.TLSCertKeyDir))
+	if err != nil {
+		return perrs.AddStack(err)
+	}
+
+	b := m.sshTaskBuilder(name, topo, base.User, options).
+		Func("StartCluster", func(ctx *task.Context) error {
+			return operator.Start(ctx, topo, options, tlsCfg)
+		})
+
+	for _, f := range fn {
+		f(b, metadata)
+	}
+
+	t := b.Build()
+
+	if err := t.Execute(task.NewContext()); err != nil {
+		if errorx.Cast(err) != nil {
+			// FIXME: Map possible task errors and give suggestions.
+			return err
+		}
+		return perrs.Trace(err)
+	}
+
+	log.Infof("Started cluster `%s` successfully", name)
+	return nil
+}
+
+// StopCluster stop the cluster.
+func (m *Manager) StopCluster(clusterName string, options operator.Options) error {
+	metadata, err := m.meta(clusterName)
+	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	tlsCfg, err := topo.TLSConfig(m.specManager.Path(clusterName, spec.TLSCertKeyDir))
+	if err != nil {
+		return err
+	}
+
+	t := m.sshTaskBuilder(clusterName, topo, base.User, options).
+		Func("StopCluster", func(ctx *task.Context) error {
+			return operator.Stop(ctx, topo, options, tlsCfg)
+		}).
+		Build()
+
+	if err := t.Execute(task.NewContext()); err != nil {
+		if errorx.Cast(err) != nil {
+			// FIXME: Map possible task errors and give suggestions.
+			return err
+		}
+		return perrs.Trace(err)
+	}
+
+	log.Infof("Stopped cluster `%s` successfully", clusterName)
+	return nil
+}
+
+// RestartCluster restart the cluster.
+func (m *Manager) RestartCluster(clusterName string, options operator.Options) error {
+	metadata, err := m.meta(clusterName)
+	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	tlsCfg, err := topo.TLSConfig(m.specManager.Path(clusterName, spec.TLSCertKeyDir))
+	if err != nil {
+		return err
+	}
+
+	t := m.sshTaskBuilder(clusterName, topo, base.User, options).
+		Func("RestartCluster", func(ctx *task.Context) error {
+			return operator.Restart(ctx, topo, options, tlsCfg)
+		}).
+		Build()
+
+	if err := t.Execute(task.NewContext()); err != nil {
+		if errorx.Cast(err) != nil {
+			// FIXME: Map possible task errors and give suggestions.
+			return err
+		}
+		return perrs.Trace(err)
+	}
+
+	log.Infof("Restarted cluster `%s` successfully", clusterName)
+	return nil
+}
