+++ pkg/cluster/manager/exec.go
@@ -1,0 +1,100 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package manager
+
+import (
+	"github.com/fatih/color"
+	"github.com/joomcode/errorx"
+	perrs "github.com/pingcap/errors"
+	operator "github.com/pingcap/tiup/pkg/cluster/operation"
+	"github.com/pingcap/tiup/pkg/cluster/spec"
+	"github.com/pingcap/tiup/pkg/cluster/task"
+	"github.com/pingcap/tiup/pkg/logger/log"
+	"github.com/pingcap/tiup/pkg/set"
+)
+
+// ExecOptions for exec shell commanm.
+type ExecOptions struct {
+	Command string
+	Sudo    bool
+}
+
+// Exec shell command on host in the tidb cluster.
+func (m *Manager) Exec(clusterName string, opt ExecOptions, gOpt operator.Options) error {
+	metadata, err := m.meta(clusterName)
+	if err != nil {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	filterRoles := set.NewStringSet(gOpt.Roles...)
+	filterNodes := set.NewStringSet(gOpt.Nodes...)
+
+	var shellTasks []task.Task
+	uniqueHosts := map[string]int{} // host -> ssh-port
+	topo.IterInstance(func(inst spec.Instance) {
+		if _, found := uniqueHosts[inst.GetHost()]; !found {
+			if len(gOpt.Roles) > 0 && !filterRoles.Exist(inst.Role()) {
+				return
+			}
+
+			if len(gOpt.Nodes) > 0 && !filterNodes.Exist(inst.GetHost()) {
+				return
+			}
+
+			uniqueHosts[inst.GetHost()] = inst.GetSSHPort()
+		}
+	})
+
+	for host := range uniqueHosts {
+		shellTasks = append(shellTasks,
+			task.NewBuilder().
+				Shell(host, opt.Command, opt.Sudo).
+				Build())
+	}
+
+	t := m.sshTaskBuilder(clusterName, topo, base.User, gOpt).
+		Parallel(false, shellTasks...).
+		Build()
+
+	execCtx := task.NewContext()
+	if err := t.Execute(execCtx); err != nil {
+		if errorx.Cast(err) != nil {
+			// FIXME: Map possible task errors and give suggestions.
+			return err
+		}
+		return perrs.Trace(err)
+	}
+
+	// print outputs
+	for host := range uniqueHosts {
+		stdout, stderr, ok := execCtx.GetOutputs(host)
+		if !ok {
+			continue
+		}
+		log.Infof("Outputs of %s on %s:",
+			color.CyanString(opt.Command),
+			color.CyanString(host))
+		if len(stdout) > 0 {
+			log.Infof("%s:\n%s", color.GreenString("stdout"), stdout)
+		}
+		if len(stderr) > 0 {
+			log.Infof("%s:\n%s", color.RedString("stderr"), stderr)
+		}
+	}
+
+	return nil
+}
